refactor: move resolve_github_mirror_url to utils/helpers
- Extract GitHub mirror URL resolution logic from
  `Config::resolve_github_mirror_url` into a standalone function
  `resolve_github_mirror_url` in `helpers.rs`
- Update `Config::resolve_github_mirror_url` to delegate to the new
  helper function
- Add necessary `use dwall::config::Network` import and function
  definition in `helpers.rs`
